Extracted Wheel class into seperate code and header files. Additionally moved 'hermite' function into the wheel class since it contained its only use.
