+++ CMakeLists.txt
@@ -45,6 +45,7 @@ add_library(${PROJECT_NAME}
   src/pos_joint.cpp
   src/vel_joint.cpp
   src/pid_joint.cpp
+  src/wheel.cpp
 )
 
 target_link_libraries(${PROJECT_NAME} ${catkin_LIBRARIES})
